Refactor: convert search screen to LazyVerticalGrid
This commit refactors the search screen to use LazyVerticalGrid for improved performance and flexibility in displaying search results.

The following changes were made:
- Replaced Column with LazyVerticalGrid in `SearchScreen.kt`.
- Extracted UI sections into separate files:
    - `RecentSearchesSection.kt`
    - `SearchScreenHeaderSection.kt`
    - `SuccessMediaItemsSection.kt`
    - `SuggestionsHubSection.kt`
- Deleted the old section files:
    - `RecentSearches.kt`
    - `SuggestionsHub.kt`
- Adjusted layout parameters and paddings to suit the grid layout.
- Updated `SearchViewModel.kt` to accommodate changes (minor logging adjustment).
