+++ ui/src/main/java/com/amsterdam/ui/screens/search/keywordSearch/sections/SearchScreenHeaderSection.kt
@@ -1,0 +1,86 @@
+package com.amsterdam.ui.screens.search.keywordSearch.sections
+
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.grid.GridItemSpan
+import androidx.compose.foundation.lazy.grid.LazyGridScope
+import androidx.compose.foundation.text.KeyboardActions
+import androidx.compose.ui.Modifier
+import com.amsterdam.designsystem.R
+import androidx.compose.ui.layout.onSizeChanged
+import androidx.compose.ui.platform.SoftwareKeyboardController
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.unit.IntSize
+import androidx.compose.ui.unit.dp
+import com.amsterdam.designsystem.components.TabsLayout
+import com.amsterdam.designsystem.components.TextField
+import com.amsterdam.designsystem.theme.AppTheme
+import com.amsterdam.ui.components.appBar.DefaultAppBar
+import com.amsterdam.viewmodel.search.keywordSearch.TabOption
+
+fun LazyGridScope.searchScreenHeaderSection(
+    keyword: String,
+    selectedTabOption: TabOption,
+    onNavigateBackClicked: () -> Unit,
+    onKeywordValuedChanged: (String) -> Unit,
+    onFilterButtonClicked: () -> Unit,
+    onTabOptionClicked: (TabOption) -> Unit,
+    onSaveSearchHistory: () -> Unit,
+    keyboardController: SoftwareKeyboardController?,
+    modifier: Modifier = Modifier,
+    onHeaderSizeChanged: (IntSize) -> Unit = {},
+) {
+    //val keyboardController = LocalSoftwareKeyboardController.current
+    item(span = { GridItemSpan(maxLineSpan) }) {
+        Column(
+            modifier = modifier.onSizeChanged(onSizeChanged = onHeaderSizeChanged),
+        ) {
+            DefaultAppBar(
+                modifier = Modifier.padding(horizontal = 16.dp),
+                title = stringResource(R.string.search),
+                onNavigateBackClicked = onNavigateBackClicked,
+            )
+            TextField(
+                modifier =
+                    Modifier
+                        .background(color = AppTheme.color.surface)
+                        .padding(top = 8.dp)
+                        .padding(horizontal = 16.dp),
+                text = keyword,
+                onValueChange = onKeywordValuedChanged,
+                hintText = stringResource(R.string.search_hint),
+                trailingIcon = com.amsterdam.ui.R.drawable.ic_filter_vertical,
+                onTrailingClick = onFilterButtonClicked,
+                isTrailingClickEnabled = keyword.isNotBlank(),
+                isError = keyword.length > 100,
+                errorMessage = stringResource(R.string.search_error_query_too_long),
+                maxCharacters = 100,
+                keyboardActions =
+                    KeyboardActions(
+                        onSearch = {
+                            keyboardController?.hide()
+                            onKeywordValuedChanged(keyword)
+                            onSaveSearchHistory()
+                        },
+                    ),
+                imeAction = ImeAction.Search,
+            )
+            AnimatedVisibility(keyword.isNotBlank()) {
+                TabsLayout(
+                    modifier = Modifier.fillMaxWidth(),
+                    tabs =
+                        listOf(
+                            stringResource(R.string.movies),
+                            stringResource(R.string.tv_shows),
+                        ),
+                    selectedIndex = selectedTabOption.index,
+                    onSelectTab = { index -> onTabOptionClicked(TabOption.entries[index]) },
+                )
+            }
+        }
+    }
+}
